Implement parsed token
