+++ md/src/lib.rs
@@ -1,5 +1,7 @@
 //pub mod elements;
 //pub mod parser;
 pub mod lexer;
+pub mod macros;
 pub mod token;
+//pub mod traits;
 pub mod errors;
